// app-native-expo-demos/app/design/type.d.ts
export interface ISystemTheme {
  radiusImageS: number;
  radiusImageM: number;
  radiusInCard: number;
  radiusCard: number;
  radiusComp1: number;
  radiusComp2: number;
  spaceElementsMin: number;
  spaceElementsXxxs: number;
  spaceElementsXxs: number;
  spaceElementsXs: number;
  spaceElementsS: number;
  spaceElementsM: number;
  spaceElementsL: number;
  spaceElementsXl: number;
  spaceCardPaddingUpdownS: number;
  spaceCardPaddingLeftRightS: number;
  spaceCardPaddingUpdownXs: number;
  spaceCardPaddingLeftRightXs: number;
  spaceCardPaddingXms: number;
  spaceCardPaddingLeftRightXxs: number;
  spaceCardPaddingLeftRightXxxs: number;
  iconPrimary: string;
  textPrimary: string;
  textTitle: string;
  textPrimaryDisabled: string;
  textInverse: string;
  textSecondary: string;
  textErrorDefault: string;
  textInfoDefault: string;
  textWarningDefault: string;
  textCautionDefault: string;
  textMethodDefault: string;
  chartBorderRain: string;
  chartBgRain: string;
  containerPrimary: string;
  containerSecondary: string;
  containerTertiary: string;
  containerFifth: string;
  containerPrimaryDisabled: string;
  containerBpDefault: string;
  containerBpPress: string;
  containerBpDisabled: string;
  containerBsDefault: string;
  containerBsPress: string;
  containerBsDisabled: string;
  containerBpWeakDefault: string;
  containerErrorWeakDefault: string;
  containerWarningWeakDefault: string;
  containerInfoWeakDefault: string;
  containerCautionWeakDefault: string;
  containerSwitchOffDisabled: string;
  dividerDefault: string;
  dividerList: string;
  dividerBp: string;
  borderWeakDefault: string;
  sizeIconXs: number;
  sizeIconXl: number;
  sizeIconXxl: number;
  sizeCompXxxs: number;
  sizeComphXxs: number;
  sizeComphXs: number;
  sizeComphS: number;
  sizeComphM: number;
  sizeComphL: number;
  sizeSwitch: number;
  sizeComphMin: number;
  sizeComphWM: number;
  sizeComphWXl: number;
  sizeComphWXxl: number;
  sizeComphWXxxl: number;
  sizeAvatarL: number;
  sizeAvatarM: number;
  cnHeadlineXxsStrong: {
    fontSize: number;
    fontWeight: string;
    lineHeight: number;
  };
  cnHeadlineXsStrong: {
    fontSize: number;
    fontWeight: string;
    lineHeight: number;
  };
  cnHeadlineMStrong: {
    fontSize: number;
    fontWeight: string;
    lineHeight: number;
  };
  cnBodyS: {
    fontSize: number;
    fontWeight: string;
    lineHeight: number;
  };
  cnBodyM: {
    fontSize: number;
    fontWeight: string;
    lineHeight: number;
  };
  cnBodyL: {
    fontSize: number;
    fontWeight: string;
    lineHeight: number;
  };
  cnDisplayLStrong: {
    fontSize: number;
    fontWeight: string;
    lineHeight: number;
  };
  cnDisplayXsStrong: {
    fontSize: number;
    fontWeight: string;
    lineHeight: number;
  };
  cnDisplayXxsStrong: {
    fontSize: number;
    fontWeight: string;
    lineHeight: number;
  };
  cnDisplayXxs: {
    fontSize: number;
    fontWeight: string;
    lineHeight: number;
  };
  cnDisplayXxxsStrong: {
    fontSize: number;
    fontWeight: string;
    lineHeight: number;
  };
  containerLinearGradient: string;
  backgroundBlur: string;
}